[base] Remove `BackdropUnstyled` component
