+++ docs/data/material/components/backdrop/backdrop.md
@@ -1,7 +1,7 @@
 ---
 product: material-ui
 title: Backdrop React Component
-components: Backdrop, BackdropUnstyled
+components: Backdrop
 githubLabel: 'component: backdrop'
 ---
 
@@ -17,12 +17,3 @@ In its simplest form, the backdrop component will add a dimmed layer over your a
 ## Example
 
 {{"demo": "SimpleBackdrop.js"}}
-
-## Unstyled
-
-The backdrop also comes with the Base package.
-It's ideal for doing heavy customizations and minimizing bundle size.
-
-```js
-import BackdropUnstyled from '@mui/base/BackdropUnstyled';
-```
