--- docs/pages/base/api/backdrop-unstyled.json
@@ -1,28 +1,0 @@
-{
-  "props": {
-    "children": { "type": { "name": "node" } },
-    "classes": { "type": { "name": "object" } },
-    "component": { "type": { "name": "elementType" } },
-    "components": {
-      "type": { "name": "shape", "description": "{ Root?: elementType }" },
-      "default": "{}"
-    },
-    "componentsProps": {
-      "type": { "name": "shape", "description": "{ root?: object }" },
-      "default": "{}"
-    },
-    "invisible": { "type": { "name": "bool" } }
-  },
-  "name": "BackdropUnstyled",
-  "styles": {
-    "classes": ["root", "invisible"],
-    "globalClasses": { "root": "MuiBackdrop-root", "invisible": "MuiBackdrop-invisible" },
-    "name": null
-  },
-  "spread": true,
-  "forwardsRefTo": "HTMLDivElement",
-  "filename": "/packages/mui-base/src/BackdropUnstyled/BackdropUnstyled.js",
-  "inheritance": null,
-  "demos": "<ul><li><a href=\"/material-ui/react-backdrop/\">Backdrop</a></li></ul>",
-  "cssComponent": false
-}
