--- docs/translations/api-docs/backdrop-unstyled/backdrop-unstyled-pt.json
@@ -1,21 +1,0 @@
-{
-  "componentDescription": "",
-  "propDescriptions": {
-    "children": "O conteúdo do componente.",
-    "classes": "Sobrescreve ou extende os estilos aplicados para o componente. Veja a <a href=\"#css\">API CSS</a> abaixo para maiores detalhes.",
-    "component": "The component used for the root node. Either a string to use a HTML element or a component.",
-    "components": "The components used for each slot inside the Backdrop. Either a string to use a HTML element or a component.",
-    "componentsProps": "The props used for each slot inside the Backdrop.",
-    "invisible": "If <code>true</code>, the backdrop is invisible. It can be used when rendering a popover or a custom select component."
-  },
-  "classDescriptions": {
-    "root": {
-      "description": "Estilos aplicados ao elemento raiz."
-    },
-    "invisible": {
-      "description": "Estilos aplicados para {{nodeName}} se {{conditions}}.",
-      "nodeName": "o elemento raiz",
-      "conditions": "<code>invisible={true}</code>"
-    }
-  }
-}
