--- docs/translations/api-docs/backdrop-unstyled/backdrop-unstyled.json
@@ -1,19 +1,0 @@
-{
-  "componentDescription": "",
-  "propDescriptions": {
-    "children": "The content of the component.",
-    "classes": "Override or extend the styles applied to the component. See <a href=\"#css\">CSS API</a> below for more details.",
-    "component": "The component used for the root node. Either a string to use a HTML element or a component.",
-    "components": "The components used for each slot inside the Backdrop. Either a string to use a HTML element or a component.",
-    "componentsProps": "The props used for each slot inside the Backdrop.",
-    "invisible": "If <code>true</code>, the backdrop is invisible. It can be used when rendering a popover or a custom select component."
-  },
-  "classDescriptions": {
-    "root": { "description": "Styles applied to the root element." },
-    "invisible": {
-      "description": "Styles applied to {{nodeName}} if {{conditions}}.",
-      "nodeName": "the root element",
-      "conditions": "<code>invisible={true}</code>"
-    }
-  }
-}
