+++ docs/translations/api-docs/modal-unstyled/modal-unstyled.json
@@ -2,7 +2,7 @@
   "componentDescription": "Modal is a lower-level construct that is leveraged by the following components:\n\n- [Dialog](/api/dialog/)\n- [Drawer](/api/drawer/)\n- [Menu](/api/menu/)\n- [Popover](/api/popover/)\n\nIf you are creating a modal dialog, you probably want to use the [Dialog](/api/dialog/) component\nrather than directly using Modal.\n\nThis component shares many concepts with [react-overlays](https://react-bootstrap.github.io/react-overlays/#modals).",
   "propDescriptions": {
     "BackdropComponent": "A backdrop component. This prop enables custom backdrop rendering.",
-    "BackdropProps": "Props applied to the <a href=\"/api/backdrop-unstyled/\"><code>BackdropUnstyled</code></a> element.",
+    "BackdropProps": "Props applied to the backdrop element.",
     "children": "A single child content element.<br>⚠️ <a href=\"/guides/composition/#caveat-with-refs\">Needs to be able to hold a ref</a>.",
     "classes": "Override or extend the styles applied to the component. See <a href=\"#css\">CSS API</a> below for more details.",
     "closeAfterTransition": "When set to true the Modal waits until a nested Transition is completed before closing.",
