--- packages/mui-base/src/BackdropUnstyled/BackdropUnstyled.test.js
@@ -1,60 +1,0 @@
-import * as React from 'react';
-import { expect } from 'chai';
-import { createRenderer, describeConformance } from 'test/utils';
-import BackdropUnstyled, { backdropUnstyledClasses as classes } from '@mui/base/BackdropUnstyled';
-
-describe('<BackdropUnstyled />', () => {
-  const { render } = createRenderer();
-
-  describeConformance(
-    <BackdropUnstyled>
-      <div />
-    </BackdropUnstyled>,
-    () => ({
-      classes,
-      inheritComponent: 'div',
-      render,
-      refInstanceof: window.HTMLDivElement,
-      testComponentPropWith: 'div',
-      skip: [
-        'themeDefaultProps', // unstyled
-        'themeStyleOverrides', // unstyled
-        'themeVariants', // unstyled
-      ],
-    }),
-  );
-
-  it('forwards style props on the Root component', () => {
-    let ownerState = null;
-    let theme = null;
-
-    const Root = React.forwardRef(
-      ({ ownerState: ownerStateProp, theme: themeProp, ...other }, ref) => {
-        ownerState = ownerStateProp;
-        theme = themeProp;
-        return <span ref={ref} {...other} />;
-      },
-    );
-
-    render(<BackdropUnstyled components={{ Root }} />);
-
-    expect(ownerState).not.to.equal(null);
-    expect(theme).not.to.equal(null);
-  });
-
-  it('does not forward style props as DOM attributes if component slot is primitive', () => {
-    const elementRef = React.createRef();
-    render(
-      <BackdropUnstyled
-        components={{
-          Root: 'span',
-        }}
-        ref={elementRef}
-      />,
-    );
-
-    const { current: element } = elementRef;
-    expect(element.getAttribute('ownerState')).to.equal(null);
-    expect(element.getAttribute('theme')).to.equal(null);
-  });
-});
