--- packages/mui-base/src/BackdropUnstyled/index.d.ts
@@ -1,5 +1,0 @@
-export { default } from './BackdropUnstyled';
-export * from './BackdropUnstyled';
-
-export { default as backdropUnstyledClasses } from './backdropUnstyledClasses';
-export * from './backdropUnstyledClasses';
