--- packages/mui-base/src/BackdropUnstyled/index.js
@@ -1,6 +1,0 @@
-export { default } from './BackdropUnstyled';
-
-export {
-  default as backdropUnstyledClasses,
-  getBackdropUtilityClass,
-} from './backdropUnstyledClasses';
