+++ packages/mui-material/src/Backdrop/Backdrop.d.ts
@@ -1,21 +1,45 @@
 import * as React from 'react';
 import { SxProps } from '@mui/system';
-import { ExtendBackdropUnstyledTypeMap, BackdropUnstyledTypeMap } from '@mui/base/BackdropUnstyled';
 import { FadeProps } from '../Fade';
 import { TransitionProps } from '../transitions/transition';
 import { Theme } from '../styles';
+import { BackdropClasses } from './backdropClasses';
 import { OverridableComponent, OverrideProps } from '../OverridableComponent';
 
-export type BackdropTypeMap<
-  D extends React.ElementType = 'span',
-  P = {},
-> = ExtendBackdropUnstyledTypeMap<{
+export interface BackdropComponentsPropsOverrides {}
+
+export interface BackdropTypeMap<P = {}, D extends React.ElementType = 'div'> {
   props: P &
     Partial<Omit<FadeProps, 'children'>> & {
       /**
+       * The content of the component.
+       */
+      children?: React.ReactNode;
+      /**
+       * The components used for each slot inside the Backdrop.
+       * Either a string to use a HTML element or a component.
+       * @default {}
+       */
+      components?: {
+        Root?: React.ElementType;
+      };
+      /**
+       * The props used for each slot inside the Backdrop.
+       * @default {}
+       */
+      componentsProps?: {
+        root?: React.HTMLAttributes<HTMLDivElement> & BackdropComponentsPropsOverrides;
+      };
+      /**
        * Override or extend the styles applied to the component.
        */
-      classes?: BackdropUnstyledTypeMap['props']['classes'];
+      classes?: Partial<BackdropClasses>;
+      /**
+       * If `true`, the backdrop is invisible.
+       * It can be used when rendering a popover or a custom select component.
+       * @default false
+       */
+      invisible?: boolean;
       /**
        * If `true`, the component is shown.
        */
@@ -31,13 +55,11 @@ export type BackdropTypeMap<
       transitionDuration?: TransitionProps['timeout'];
     };
   defaultComponent: D;
-}>;
+}
 
 type BackdropRootProps = NonNullable<BackdropTypeMap['props']['componentsProps']>['root'];
 
 export const BackdropRoot: React.FC<BackdropRootProps>;
-
-export type BackdropClassKey = keyof NonNullable<BackdropTypeMap['props']['classes']>;
 
 /**
  *
@@ -50,16 +72,11 @@ export type BackdropClassKey = keyof NonNullable<BackdropTypeMap['props']['class
  * - [Backdrop API](https://mui.com/api/backdrop/)
  * - inherits [Fade API](https://mui.com/api/fade/)
  */
-
 declare const Backdrop: OverridableComponent<BackdropTypeMap>;
-
-export type BackdropClasses = Record<BackdropClassKey, string>;
-
-export const backdropClasses: BackdropClasses;
 
 export type BackdropProps<
   D extends React.ElementType = BackdropTypeMap['defaultComponent'],
   P = {},
-> = OverrideProps<BackdropTypeMap<D, P>, D>;
+> = OverrideProps<BackdropTypeMap<P, D>, D>;
 
 export default Backdrop;
