+++ packages/mui-material/src/Backdrop/backdropClasses.ts
@@ -1,0 +1,21 @@
+import { generateUtilityClass, generateUtilityClasses } from '@mui/base';
+
+export interface BackdropClasses {
+  /** Styles applied to the root element. */
+  root: string;
+  /** Styles applied to the root element if `invisible={true}`. */
+  invisible: string;
+}
+
+export type BackdropClassKey = keyof BackdropClasses;
+
+export function getBackdropUtilityClass(slot: string): string {
+  return generateUtilityClass('MuiBackdrop', slot);
+}
+
+const backdropClasses: BackdropClasses = generateUtilityClasses('MuiBackdrop', [
+  'root',
+  'invisible',
+]);
+
+export default backdropClasses;
