+++ packages/mui-material/src/Backdrop/index.d.ts
@@ -1,2 +1,5 @@
 export { default } from './Backdrop';
 export * from './Backdrop';
+
+export { default as tableClasses } from './backdropClasses';
+export * from './backdropClasses';
