+++ packages/mui-material/src/Backdrop/index.js
@@ -1,2 +1,5 @@
 export { default } from './Backdrop';
 export * from './Backdrop';
+
+export { default as backdropClasses } from './backdropClasses';
+export * from './backdropClasses';
